Retrieval-Augmented-Generation/05_Rag_with_Weviate_and_Mistral_AI/pyproject.toml
[project]
name = "05-rag-with-weviate-and-mistral-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "rapidocr-onnxruntime>=1.4.4",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
    "weaviate-client>=4.15.3",
]
